refactor(client): decompose Minimap into renderer and interaction hook
Extract 1,813-line canvas minimap into focused modules:
- MinimapRenderer.ts (937 LOC): pure canvas 2D drawing functions
  with zero React dependencies — terrain, roads, buildings, entity
  pips, destination marker, location icons with flyweight cache
- useMinimapInteraction.ts (389 LOC): zoom, collapse, resize drag,
  click-to-teleport interaction handlers

Minimap.tsx reduced to 667 lines as orchestrator managing canvas ref,
requestAnimationFrame loop, and composing satellite hooks. Zero
behavior changes — all drawing, caching, and interaction preserved.
